Constructors and tests - Owner and Projects

+++ GaslightVillage/src/IProject.java
@@ -2,5 +2,5 @@ import Helpers.Status;
 
 public interface IProject {
   public void setApprovalStatus(Status status);
-  public Float calculateAssessment(IAccount account);
+  public Double calculateAssessment(IAccount account);
 }
